Visual overhaul: AI-forward professional design system
Transform 143.dev from generic defaults to polished, premium aesthetic matching Vercel/Linear/Cursor. Key changes: blue-tinted dark mode (`oklch(0.13 0.008 260)`), gradient primary buttons with glow effects, card hover shadows/glows, sidebar gradient wash, radial gradient background, more dramatic typography hierarchy. Updated theme colors for deeper saturation, added CSS variables for gradients and glow effects, enhanced transitions to support shadow/transform animations. Installed and integrated Geist font (sans + mono) for modern, professional typography. Updated 14 files across globals, components, pages, and UI library.
