+++ frontend/src/app/globals.css
@@ -4,11 +4,11 @@
 @custom-variant dark (&:is(.dark *));
 
 :root {
-  --border: oklch(0.92 0.005 250);
+  --border: oklch(0.91 0.004 250);
   --muted: oklch(0.955 0 0);
   --card: oklch(1 0 0);
   --radius: 0.5rem;
-  --background: oklch(0.985 0 0);
+  --background: oklch(0.98 0.003 260);
   --foreground: oklch(0.145 0 0);
   --card-foreground: oklch(0.145 0 0);
   --popover: oklch(1 0 0);
@@ -28,7 +28,7 @@
   --chart-3: oklch(0.398 0.07 227.392);
   --chart-4: oklch(0.828 0.189 84.429);
   --chart-5: oklch(0.769 0.188 70.08);
-  --sidebar: oklch(0.965 0 0);
+  --sidebar: oklch(0.97 0.005 260);
   --sidebar-foreground: oklch(0.145 0 0);
   --sidebar-primary: oklch(0.56 0.17 255);
   --sidebar-primary-foreground: oklch(0.985 0 0);
@@ -36,6 +36,10 @@
   --sidebar-accent-foreground: oklch(0.205 0 0);
   --sidebar-border: oklch(0.922 0 0);
   --sidebar-ring: oklch(0.55 0.05 250);
+  --gradient-primary: linear-gradient(135deg, oklch(0.5 0.16 270), oklch(0.45 0.18 300));
+  --gradient-primary-hover: linear-gradient(135deg, oklch(0.45 0.16 270), oklch(0.4 0.18 300));
+  --glow-primary: 0 0 20px oklch(0.5 0.16 270 / 15%);
+  --glow-primary-sm: 0 0 10px oklch(0.5 0.16 270 / 10%);
 }
 
 @theme inline {
@@ -82,22 +86,22 @@
 }
 
 .dark {
-  --background: oklch(0.145 0 0);
+  --background: oklch(0.13 0.008 260);
   --foreground: oklch(0.985 0 0);
-  --card: oklch(0.205 0 0);
+  --card: oklch(0.17 0.006 260);
   --card-foreground: oklch(0.985 0 0);
-  --popover: oklch(0.205 0 0);
+  --popover: oklch(0.17 0.006 260);
   --popover-foreground: oklch(0.985 0 0);
-  --primary: oklch(0.75 0.05 250);
+  --primary: oklch(0.7 0.15 270);
   --primary-foreground: oklch(0.145 0 0);
-  --secondary: oklch(0.269 0 0);
+  --secondary: oklch(0.22 0.005 260);
   --secondary-foreground: oklch(0.985 0 0);
-  --muted: oklch(0.269 0 0);
+  --muted: oklch(0.22 0.005 260);
   --muted-foreground: oklch(0.708 0 0);
   --accent: oklch(0.25 0.02 250);
   --accent-foreground: oklch(0.8 0.04 250);
   --destructive: oklch(0.704 0.191 22.216);
-  --border: oklch(1 0 0 / 10%);
+  --border: oklch(1 0 0 / 7%);
   --input: oklch(1 0 0 / 15%);
   --ring: oklch(0.55 0.05 250);
   --chart-1: oklch(0.65 0.06 250);
@@ -105,14 +109,18 @@
   --chart-3: oklch(0.769 0.188 70.08);
   --chart-4: oklch(0.627 0.265 303.9);
   --chart-5: oklch(0.645 0.246 16.439);
-  --sidebar: oklch(0.205 0 0);
+  --sidebar: oklch(0.11 0.01 260);
   --sidebar-foreground: oklch(0.985 0 0);
   --sidebar-primary: oklch(0.75 0.05 250);
   --sidebar-primary-foreground: oklch(0.145 0 0);
   --sidebar-accent: oklch(0.25 0.02 250);
   --sidebar-accent-foreground: oklch(0.8 0.04 250);
-  --sidebar-border: oklch(1 0 0 / 10%);
+  --sidebar-border: oklch(1 0 0 / 6%);
   --sidebar-ring: oklch(0.55 0.05 250);
+  --gradient-primary: linear-gradient(135deg, oklch(0.6 0.18 270), oklch(0.55 0.2 300));
+  --gradient-primary-hover: linear-gradient(135deg, oklch(0.55 0.18 270), oklch(0.5 0.2 300));
+  --glow-primary: 0 0 20px oklch(0.6 0.18 270 / 25%);
+  --glow-primary-sm: 0 0 10px oklch(0.6 0.18 270 / 15%);
 }
 
 @layer base {
@@ -123,9 +131,12 @@
     @apply bg-background text-foreground text-[13px] antialiased;
   }
   button, a, [role="button"] {
-    @apply transition-colors duration-150;
+    @apply transition-all duration-150;
   }
   [data-slot="card"] {
-    @apply transition-shadow duration-200;
+    @apply transition-all duration-200;
+  }
+  h1, h2, h3 {
+    @apply tracking-tight;
   }
 }
