+++ frontend/src/components/page-header.tsx
@@ -10,8 +10,8 @@ export function PageHeader({ title, description, action }: PageHeaderProps) {
   return (
     <div className="flex flex-col gap-2 sm:flex-row sm:items-start sm:justify-between">
       <div>
-        <h1 className="text-xl font-semibold tracking-tight text-foreground">{title}</h1>
-        <p className="mt-1 text-sm text-muted-foreground">{description}</p>
+        <h1 className="text-2xl font-bold tracking-tight text-foreground">{title}</h1>
+        <p className="mt-1.5 text-sm text-muted-foreground/80">{description}</p>
       </div>
       {action && <div className="shrink-0">{action}</div>}
     </div>
