+++ frontend/src/components/pm/pm-status-banner.tsx
@@ -74,7 +74,7 @@ export function PMStatusBanner({ hasActivePlanSession }: PMStatusBannerProps) {
       <div
         className={`flex items-center gap-3 rounded-lg border px-4 py-2.5 transition-colors ${
           isRunning
-            ? "border-primary/20 bg-primary/5 dark:border-primary/30 dark:bg-primary/10"
+            ? "border-primary/20 bg-primary/5 dark:border-primary/30 dark:bg-primary/10 dark:shadow-[0_0_20px_oklch(0.6_0.15_270_/_8%)]"
             : "border-border bg-muted/30"
         }`}
       >
@@ -83,7 +83,7 @@ export function PMStatusBanner({ hasActivePlanSession }: PMStatusBannerProps) {
         <span className="text-[13px] font-medium text-foreground">PM Agent</span>
 
         <span className={`text-[11px] font-medium px-1.5 py-0.5 rounded ${
-          isRunning ? "bg-primary/10 text-primary"
+          isRunning ? "bg-primary/10 text-primary shadow-[var(--glow-primary-sm)]"
           : agentStatus === "completed" ? "bg-green-100 text-green-700 dark:bg-green-900/40 dark:text-green-300"
           : agentStatus === "failed" ? "bg-destructive/10 text-destructive"
           : "bg-muted text-muted-foreground"
