+++ frontend/src/components/ui/badge.tsx
@@ -9,7 +9,7 @@ const badgeVariants = cva(
   {
     variants: {
       variant: {
-        default: "bg-primary text-primary-foreground [a&]:hover:bg-primary/90",
+        default: "bg-primary bg-[image:var(--gradient-primary)] text-white [a&]:hover:bg-primary/90",
         secondary:
           "bg-secondary text-secondary-foreground border-border/50 [a&]:hover:bg-secondary/90",
         destructive:
@@ -18,6 +18,8 @@ const badgeVariants = cva(
           "border-border text-foreground [a&]:hover:bg-accent [a&]:hover:text-accent-foreground",
         ghost: "[a&]:hover:bg-accent [a&]:hover:text-accent-foreground",
         link: "text-primary underline-offset-4 [a&]:hover:underline",
+        glow: "bg-primary/15 text-primary border-primary/20 shadow-[var(--glow-primary-sm)]",
+        success: "bg-emerald-500/15 text-emerald-600 dark:text-emerald-400 border-emerald-500/20",
       },
     },
     defaultVariants: {
